app server启动公共方法
